# Optimizing resource related code.