## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Avalonia has too many interfaces and casts are used too often which leads to degraded performance caused by this.
This PR is tweaking a few resource related interface usages to get rid of casts and speed up hot paths. 

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
|        Method |     Mean |   Error |  StdDev |   Gen 0 |   Gen 1 | Allocated |
|-------------- |---------:|--------:|--------:|--------:|--------:|----------:|
| CreateTextBox | 625.6 us | 2.25 us | 1.88 us | 90.8203 | 30.2734 |    230 KB |
| CreateCalendar | 13.67 ms | 0.121 ms | 0.113 ms | 1500.0000 | 531.2500 | 171.8750 |      4 MB |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
|        Method |     Mean |   Error |  StdDev |   Gen 0 |   Gen 1 | Allocated |
|-------------- |---------:|--------:|--------:|--------:|--------:|----------:|
| CreateTextBox | 526.3 us | 2.40 us | 1.87 us | 93.7500 | 28.3203 |    222 KB |
| CreateCalendar | 12.46 ms | 0.088 ms | 0.082 ms | 1453.1250 | 515.6250 | 156.2500 |      4 MB |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
Also since benchmark was using default theme I've changed data validation to use compiled bindings for attached properties since reflection path is really slow.

## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Documentation with user documentation

## Breaking changes
<!--- List any breaking changes here. When the PR is merged please add an entry to https://github.com/AvaloniaUI/Avalonia/wiki/Breaking-Changes -->

## Obsoletions / Deprecations
<!--- Obsolete and Deprecated attributes on APIs MUST only be included when discussed with Core team. @grokys, @kekekeks & @danwalmsley -->

## Fixed issues
<!--- If the pull request fixes issue(s) list them like this: 
Fixes #123
Fixes #456
-->