@timunie 2022-05-03
Hi @MarchingCube 

I am not sure if this is related, but I have once learned that `foreach` may be noticeable slower that `for` in many cases. See: https://www.c-sharpcorner.com/article/c-sharp-performance-of-code-for-loop-vs-for-each-loop/ . 

I leave this link here as a suggestion to test / benchmark if you think it's useful. If not, just ignore. 

Happy coding
Tim