@MarchingCube 2022-05-03
@timunie I had a good chuckle reading that article, although then I got a bit sad since people look at such content and then "optimize" their applications causing perf regressions.

For people actually wondering about `for` and `foreach` performance:
- It heavily depends on collection type and local information. For example for `Span<T>` quite often code generation for `foreach` is really good due to runtime optimization. Such optimization also depends on what compiler sees so it can devirtualize or inline code.
- Use proper benchmarking suite like `BenchmarkDotNet`, measuring perf using stopwatch is generally pointless. In that article you can probably swap `for` and `foreach` tests locations and get the opposite result :) Code has to be compiled, caches warmed up and benchmarking suites do that for you.
- Modern runtimes quite often use struct based enumerators that do not require allocations (for example `List<T>.Enumerator`)
- Enumerators sometimes do more work, for example `List<T>.Enumerator` checks if collection was not modified.
- For arrays and spans you can do tricks to avoid generation of index validity checks since that is also expensive if you have a lot of data to go through

Better article that actually uses proper way of benchmarking perf: https://www.meziantou.net/fastest-way-to-enumerate-a-list-t.htm

And this is why I've already optimized a few `for` loops or simply implemented better enumerators in certain cases, but it all had to be backed by benchmarks since runtime is a complex beast.