@avaloniaui-team 2022-05-03
You can test this PR using the following package version. `0.10.999-cibuild0020193-beta`. (feed url: https://nuget.avaloniaui.net/repository/avalonia-all/index.json) [PRBUILDID]